# Metal conv tensor cores