@tinyb0t 2023-08-27
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/codegen/linearizer.py            11               5
tinygrad/codegen/optimizer.py             26               7
------------------------------------------------------------
total                                     37              12
------------------------------------------------------------
lines added in the tinygrad folder: 25
```